Moving call and arguments to separate file, and added tests for call
